Fix to LHE in/out ratio (1:160)
Blood, I'm submitting this just for you to look at.  
I think your useWater method should work just fine but I've had no luck with it.  See if this is an acceptable workaround.

This does NOT investigate whether total output numbers are correct ( I don't know if they are or not ).

Sorry about...
{
  how I do brackets;
}