Fix Advanced Miner II occasionally outputting the wrong blocks
AdvMinerII would occasionally output the drops from a block recently
broken by the player. This happens when the wrong coordinates are
passed to GT_Block_Ores_Abstract.getDrops(), which causes it to fail to
find the TileEntity and then fall back to whatever the most recently
broken block was.

Fixes #751 (probably), #674, #508