Change Oil Cracker tooltip to match reality (and vice versa)
Oil Cracker only allows Maint/Energy hatches in the middle ring. Input and output hatches can be anywhere on the sides.

Minimum casing count is changed to 18 to match the tooltip, rather than increase the tooltip to 19. This allows two input and output hatches on either side, for when you want to run two cracking recipes with the same cracker.

(It turns out the cracker doesn't care which hatch the inputs and hydro/steam are put into. I don't really care either!)