Fixed a sneaky and game-breaking bug that caused quite a lot of fluids to no longer have recipes for filling cells.
The reason for the bug was that the bit flag 256 is already in use to
designate something as empty.
Since only bit flags up to 128 were documented I had unwittingly used
the 256 bit flag for fluid creation.
The documentation now informs about the function of 256.
Automatic fluid and gas creation are now controlled via booleans.

Big thanks to Techlone for helping me debug.