Made rubber consistent, rewrote WireProcessing, implemented ht cables
All rubbers are now made by first producing the polymer and then
vulcanizing it with Sulfur (equivalent to natural rubber).

Rewrote WireProcessing to reduce the amount of duplicate code.

Implemented recipes that utilize a combination of a dielectric and a
synthetic rubber to make cables.
This halves total insulation cost for below-EV cables and reduces it to
75% for EV+ cables (conductive polymer not yet implemented).