Fixed Silicone Rubber recipe, nerfed Silicone Rubber, added DT desc.
Made it so that GT_Values.NI and GT_Values.NF are consistently used when
applicable.

Nerfed Silicone Rubber so that you only get 3/7 as much as from a
regular plastic.
The idea is that Silicone Rubber is more versatile than PVC or SBR, but
also more expensive.

Changed the Distillation Tower description so that it aligns with the
new behavior.