Moved the Casings for the Large Chemical Reactor again.
As it turns out overwriting Casings 4.4 and 4.5 was not a good idea
because even though there is no actual Casing at those IDs, the textures
are in use.

I started a new batch of Casings (GT_Block_Casings8) at IDs 112-127.
I did not use IDs 80-95 because there seem to already be some textures
there.
I did not use IDs 96-111 because they seem to already be in use by
Technus.