Greatly increased IC2 Biomass consumption, added scaleable FB recipes
Previously the energy density of IC2 Biomass was very high.
1000L of IC2 biomass could be converted into 32000L of IC2 Biogas in the
Distillery or into 64000L of IC2 Biogas in the Distillation Tower.

With an IC2 Canning Machine and a Distillery you got 128000 EU from 1
Potato.
With a Pyrolyse Oven and a Distillation Tower you got 320000 EU from 1
Potato.

With the increased consumption rates (x16/x32) of Biomass the energy
Density has been greatly reduced.
1 Potato now yields 8000 EU with an IC2 Canning Machine and 12000 EU
with a Pyrolyse Oven.

Also, scaleable Recipes for Fermented Biomass has been added.
It is now possible to produce Fermented Biomass in the Pyrolyse Oven
without the need of any Fermenters.