Oil Cracker can now also works right -> left, EU/Maint Hatches anywhere
The Oil Cracker now also accepts Input Hatches on the right side and
Output Hatches on the left side.
Input/Output Hatches must be on opposite sides of one another.

Also, Energy/Maintenance Hatches no longer need to be on the middle ring
but can also be put on the sides.