Fix transformers not exploding when receiving overvoltage through P2P-EU tunnels
Transformers were using the IC2 compat to draw IC2 energy from the
tunnel before the tunnel got a chance to push GT energy into the
transformer. Because this uses up the transformer's max amperage, the
tunnel never gets a chance to push the GT energy into it.

This change makes sure the IC2 compat is not activated for P2P-EU
tunnels.