CLS support
OpenGL is a bit of a pita, wants only to be used in main thread. The easiest solution was to trigger CLS screen updates from GT. It is a soft dependency, if CLS isn't present, nothing breaks. Second pr because cloned fresh repo or else it attributed not my changes to me