isRecipeInputEqual optimization (#351)
* Optimize GT_Recipe.isRecipeInputEqual

Used HEAVILY to catch duplicate recipes during game startup. Worst-case, it
would result in 5 GT_OreDictUnificator.get calls for each pair of inputs.
Instead, this is closer to 1 + 1/N, and as an added bonus, the modified
unification it now uses doesn't copy its output only for it to be discarded
immediately after comparison.

Apart from inproved startup times, this might make a small difference when
machines switch recipes.

I am making an assumption that
GT_OreDictUnificator.get(true, GT_OreDictUnificator.get(false, stack))
is equivalent to a just GT_OreDictUnificator.get(false, stack), and the
original code only double-unificated because it was easier to code. If this
turns out to be false, there's still some performance gain from factoring
out the tStack unification, but some of the optimization here would be lost.

* Changed method visibility according to PR feedback