Tweak rutile and magnesium chloride recipe time
This makes the titanium foundry a little easier to build, as now the chemical reactor processing rutile and magnesium chloride can reach a 1:1 ratio, which is much easier to manage than the previous 5:3 ratio. Magnesium chloride recipe EU/t is increased to keep the overall EU consumption the same (312000EU).